added TaxCalculator class to consolidate calcTax methods
