+++ src/module10/Books.java
@@ -10,16 +10,4 @@ public abstract class Books extends Goods
         this.setTax(Books.tax);
     }
 
-    public static double calcTax(double tax, double price, boolean isImport)
-    {
-        double totalItemTax;
-
-        if (isImport) {
-            totalItemTax = (tax + Imports.IMPORT_TAX) * price;
-        } else {
-            totalItemTax = tax * price;
-        }
-        return Math.round(totalItemTax * 20.0) / 20.0;
-    }
-
 }
