+++ src/module10/Goods.java
@@ -13,18 +13,7 @@ public class Goods
         return tax;
     }
 
-    public void setTax(double tax) { }
-
-    public static double calcTax(double tax, double price, boolean isImport)
+    protected void setTax(double tax)
     {
-        double totalItemTax;
-
-        if (isImport) {
-            totalItemTax = (tax + Imports.IMPORT_TAX) * price;
-        } else {
-            totalItemTax = tax * price;
-        }
-        return Math.round(totalItemTax * 20.0) / 20.0;
     }
-
 }
