+++ src/module10/Purchases.java
@@ -12,13 +12,13 @@ public class Purchases
 
         // Input/Output 1
         System.out.println("Output 1");
-        double purchaseTax1a = Books.calcTax(Books.tax, Book1249.price, Book1249.isImport);
+        double purchaseTax1a = TaxCalculator.calcTax(Books.tax, Book1249.price, Book1249.isImport);
         System.out.println(Book1249.name + ": " + Book1249.price + " plus tax: " + purchaseTax1a + " = " + money.format(Book1249.price + purchaseTax1a));
 
-        double purchaseTax1b = Goods.calcTax(MusicCD1499.tax, MusicCD1499.price, MusicCD1499.isImport);
+        double purchaseTax1b = TaxCalculator.calcTax(MusicCD1499.tax, MusicCD1499.price, MusicCD1499.isImport);
         System.out.println(MusicCD1499.name + ": " + MusicCD1499.price + " plus tax: " + purchaseTax1b + " = " + money.format(MusicCD1499.price + purchaseTax1b));
 
-        double purchaseTax1c = Food.calcTax(ChocolateBar085.tax, ChocolateBar085.price, ChocolateBar085.isImport);
+        double purchaseTax1c = TaxCalculator.calcTax(ChocolateBar085.tax, ChocolateBar085.price, ChocolateBar085.isImport);
         System.out.println(ChocolateBar085.name + ": " + ChocolateBar085.price + " plus tax: " + purchaseTax1c + " = " + money.format(ChocolateBar085.price + purchaseTax1c));
 
         double outputTax1 = (purchaseTax1a + purchaseTax1b + purchaseTax1c);
@@ -33,10 +33,10 @@ public class Purchases
 
         // Input/Output 2
         System.out.println("Output 2");
-        double purchaseTax2a = Food.calcTax(ImportedBoxOfChocolates1000.tax, ImportedBoxOfChocolates1000.price, ImportedBoxOfChocolates1000.isImport);
+        double purchaseTax2a = TaxCalculator.calcTax(ImportedBoxOfChocolates1000.tax, ImportedBoxOfChocolates1000.price, ImportedBoxOfChocolates1000.isImport);
         System.out.println(ImportedBoxOfChocolates1000.name + ": " + ImportedBoxOfChocolates1000.price + " plus tax: " + purchaseTax2a + " = " + money.format(ImportedBoxOfChocolates1000.price + purchaseTax2a));
 
-        double purchaseTax2b = Goods.calcTax(ImportedBottleOfPerfume4750.tax, ImportedBottleOfPerfume4750.price, ImportedBottleOfPerfume4750.isImport);
+        double purchaseTax2b = TaxCalculator.calcTax(ImportedBottleOfPerfume4750.tax, ImportedBottleOfPerfume4750.price, ImportedBottleOfPerfume4750.isImport);
         System.out.println(ImportedBottleOfPerfume4750.name + ": " + ImportedBottleOfPerfume4750.price + " plus tax: " + purchaseTax2b + " = " + money.format(ImportedBottleOfPerfume4750.price + purchaseTax2b));
 
         double outputTax2 = purchaseTax2a + purchaseTax2b;
@@ -51,16 +51,16 @@ public class Purchases
 
         // Input/Output 3
         System.out.println("Output 3");
-        double purchaseTax3a = Goods.calcTax(ImportedBottleOfPerfume2799.tax, ImportedBottleOfPerfume2799.price, ImportedBottleOfPerfume2799.isImport);
+        double purchaseTax3a = TaxCalculator.calcTax(ImportedBottleOfPerfume2799.tax, ImportedBottleOfPerfume2799.price, ImportedBottleOfPerfume2799.isImport);
         System.out.println(ImportedBottleOfPerfume2799.name + ": " + ImportedBottleOfPerfume2799.price + " plus tax: " + purchaseTax3a + " = " + money.format(ImportedBottleOfPerfume2799.price + purchaseTax3a));
 
-        double purchaseTax3b = Goods.calcTax(BottleOfPerfume1899.tax, BottleOfPerfume1899.price, BottleOfPerfume1899.isImport);
+        double purchaseTax3b = TaxCalculator.calcTax(BottleOfPerfume1899.tax, BottleOfPerfume1899.price, BottleOfPerfume1899.isImport);
         System.out.println(BottleOfPerfume1899.name + ": " + BottleOfPerfume1899.price + " plus tax: " + purchaseTax3b + " = " + money.format(BottleOfPerfume1899.price + purchaseTax3b));
 
-        double purchaseTax3c = MedicalProducts.calcTax(PacketOfHeadachePills975.tax, PacketOfHeadachePills975.price, PacketOfHeadachePills975.isImport);
+        double purchaseTax3c = TaxCalculator.calcTax(PacketOfHeadachePills975.tax, PacketOfHeadachePills975.price, PacketOfHeadachePills975.isImport);
         System.out.println(PacketOfHeadachePills975.name + ": " + PacketOfHeadachePills975.price + " plus tax: " + purchaseTax3c + " = " + money.format(PacketOfHeadachePills975.price + purchaseTax3c));
 
-        double purchaseTax3d = Food.calcTax(BoxOfImportedChocolates1125.tax, BoxOfImportedChocolates1125.price, BoxOfImportedChocolates1125.isImport);
+        double purchaseTax3d = TaxCalculator.calcTax(BoxOfImportedChocolates1125.tax, BoxOfImportedChocolates1125.price, BoxOfImportedChocolates1125.isImport);
         System.out.println(BoxOfImportedChocolates1125.name + ": " + BoxOfImportedChocolates1125.price + " tax: " + purchaseTax3d + " = " + money.format(BoxOfImportedChocolates1125.price + purchaseTax3d));
 
         double outputTax3 = purchaseTax3a + purchaseTax3b + purchaseTax3c + purchaseTax3d;
